(refactor) Utility functions
